+++ src/styles/components/_createStudy.scss
