+++ src/styles/main.scss
@@ -36,7 +36,7 @@
 @import 'components/todoList';
 @import 'components/input';
 @import 'components/keyword';
-@import 'components/createStudyForm';
+@import 'components/createStudy';
 @import 'components/calendarPeriod';
 
 /* 레이아웃 */
